Add Misty's yes-no box to all Gym trainers
For Misty's script, I added a yes-no prompt to disable Gym trainers or
keep them enabled (since Miki would want to fight them all but casual
players might not). This commit documents it a bit more and adds it to
all other Gym scripts.

In the future, it's just a change in the call command to go to 0x800000
or 0x8800000.